test: broken query tests into multiple files
